added H3 uber generate heatmap function.
changed longitude and latitude from int to double

+++ backend-konijn/src/main/java/nl/cfns/basicpojo/MeasurementPoint.java
@@ -16,7 +16,7 @@ public class MeasurementPoint {
 	private double y;
 	private double value;
 
-	public MeasurementPoint(Integer latitude, Integer longitude, Integer RSSI) {
+	public MeasurementPoint(Double latitude, Double longitude, Integer RSSI) {
 		this.id = null;
 		this.x = latitude;
 		this.y = longitude;
